refactor: remove circular deps (#4214)
