+++ packages/blocks/src/__internal__/content-parser/index.ts
@@ -11,6 +11,7 @@ import { toast } from '../../components/toast.js';
 import type { PageBlockModel } from '../../models.js';
 import type { EdgelessPageBlockComponent } from '../../page-block/edgeless/edgeless-page-block.js';
 import { xywhArrayToObject } from '../../page-block/edgeless/utils/convert.js';
+import { registerAllBlocks } from '../service/legacy-services/index.js';
 import {
   getBlockElementById,
   getEditorContainer,
@@ -65,6 +66,7 @@ export class ContentParser {
       tableTitleColumnHandler?: TableTitleColumnHandler;
     } = {}
   ) {
+    registerAllBlocks();
     this._page = page;
     this._imageProxyEndpoint = options?.imageProxyEndpoint;
     // FIXME: this hard-coded config should be removed, see https://github.com/toeverything/blocksuite/issues/3506
