+++ packages/blocks/src/__internal__/service/legacy-services/data-view-service.ts
@@ -1,0 +1,8 @@
+import type { DataViewBlockModel } from '../../../data-view-block/data-view-model.js';
+import { BaseService } from '../service.js';
+
+export class DataViewBlockService extends BaseService<DataViewBlockModel> {
+  constructor() {
+    super();
+  }
+}
