+++ packages/blocks/src/__internal__/service/legacy-services/database-service.ts
@@ -1,0 +1,95 @@
+import { assertExists } from '@blocksuite/global/utils';
+import type { BaseBlockModel, Page } from '@blocksuite/store';
+
+import type {
+  DataViewDataType,
+  DataViewTypes,
+} from '../../../database-block/common/data-view.js';
+import type {
+  Cell,
+  Column,
+  DatabaseBlockModel,
+} from '../../../database-block/index.js';
+import type { BlockModels, SerializedBlock } from '../../utils/index.js';
+import { BaseService } from '../service.js';
+import { getService } from '../singleton.js';
+
+export class LegacyDatabaseBlockService extends BaseService<DatabaseBlockModel> {
+  initDatabaseBlock(
+    page: Page,
+    model: BaseBlockModel,
+    databaseId: string,
+    viewType: DataViewTypes,
+    isAppendNewRow = true
+  ) {
+    const blockModel = page.getBlockById(databaseId) as DatabaseBlockModel;
+    assertExists(blockModel);
+    blockModel.initTemplate(viewType);
+    if (isAppendNewRow) {
+      // Add a paragraph after database
+      const parent = page.getParent(model);
+      assertExists(parent);
+      page.addBlock('affine:paragraph', {}, parent.id);
+    }
+    blockModel.applyColumnUpdate();
+  }
+
+  override block2Json(block: BlockModels['affine:database']): SerializedBlock {
+    const columns = [...block.columns];
+    const rowIds = block.children.map(child => child.id);
+
+    const children = block.children?.map(child => {
+      return getService(child.flavour).block2Json(child);
+    });
+
+    return {
+      flavour: block.flavour,
+      databaseProps: {
+        id: block.id,
+        title: block.title.toString(),
+        rowIds,
+        cells: block.cells,
+        columns,
+        views: block.views,
+      },
+      children,
+    };
+  }
+
+  override async onBlockPasted(
+    model: BlockModels['affine:database'],
+    props: {
+      rowIds: string[];
+      columns: Column[];
+      cells: Record<string, Record<string, Cell>>;
+      views: DataViewDataType[];
+    }
+  ) {
+    const { rowIds, columns, cells, views } = props;
+    const columnIds = columns.map(column => column.id);
+    model.deleteColumn(model.id);
+    const newColumnIds = columns.map(schema => {
+      const { id, ...nonIdProps } = schema;
+      return model.addColumn('end', nonIdProps);
+    });
+    model.applyColumnUpdate();
+
+    const newRowIds = model.children.map(child => child.id);
+    rowIds.forEach((rowId, rowIndex) => {
+      const newRowId = newRowIds[rowIndex];
+      columnIds.forEach((columnId, columnIndex) => {
+        const cellData = cells[rowId]?.[columnId];
+        const value = cellData?.value;
+        if (!value) return;
+        model.updateCell(newRowId, {
+          columnId: newColumnIds[columnIndex],
+          value,
+        });
+      });
+    });
+
+    views.forEach(view => {
+      model.addView(view.mode);
+    });
+  }
+}
