+++ packages/blocks/src/__internal__/service/legacy-services/divider-service.ts
@@ -1,5 +1,5 @@
-import { BaseService } from '../__internal__/service/service.js';
-import type { DividerBlockModel } from './divider-model.js';
+import type { DividerBlockModel } from '../../../divider-block/divider-model.js';
+import { BaseService } from '../service.js';
 
 export class DividerBlockService extends BaseService<DividerBlockModel> {
   override async block2html(): Promise<string> {
