+++ packages/blocks/src/code-block/index.ts
@@ -1,4 +1,3 @@
 export * from './code-block.js';
 export * from './code-model.js';
-export * from './code-service.js';
 export * from './components/index.js';
