--- packages/blocks/src/data-view-block/data-view-service.ts
@@ -1,8 +1,0 @@
-import { BaseService } from '../__internal__/service/service.js';
-import type { DataViewBlockModel } from './data-view-model.js';
-
-export class DataViewBlockService extends BaseService<DataViewBlockModel> {
-  constructor() {
-    super();
-  }
-}
