+++ packages/blocks/src/divider-block/divider-block.ts
@@ -4,9 +4,7 @@ import { css, html } from 'lit';
 import { customElement } from 'lit/decorators.js';
 
 import { BLOCK_CHILDREN_CONTAINER_PADDING_LEFT } from '../__internal__/consts.js';
-import { registerService } from '../__internal__/service/index.js';
 import type { DividerBlockModel } from './divider-model.js';
-import { DividerBlockService } from './divider-service.js';
 
 @customElement('affine-divider')
 export class DividerBlockComponent extends BlockElement<DividerBlockModel> {
@@ -30,7 +28,6 @@ export class DividerBlockComponent extends BlockElement<DividerBlockModel> {
 
   override connectedCallback() {
     super.connectedCallback();
-    registerService('affine:divider', DividerBlockService);
   }
 
   override firstUpdated() {
