+++ packages/blocks/src/divider-block/index.ts
@@ -1,3 +1,2 @@
 export * from './divider-block.js';
 export * from './divider-model.js';
-export * from './divider-service.js';
