+++ packages/blocks/src/list-block/index.ts
@@ -1,3 +1,2 @@
 export * from './list-block.js';
 export * from './list-model.js';
-export * from './list-service.js';
