+++ packages/blocks/src/note-block/index.ts
@@ -1,3 +1,2 @@
 export * from './note-block.js';
 export * from './note-model.js';
-export * from './note-service.js';
