+++ packages/blocks/src/note-block/note-block.ts
@@ -3,10 +3,8 @@ import { BlockElement } from '@blocksuite/lit';
 import { css, html } from 'lit';
 import { customElement } from 'lit/decorators.js';
 
-import { registerService } from '../__internal__/service/index.js';
 import { bindHotKey } from './keymap.js';
 import type { NoteBlockModel } from './note-model.js';
-import { NoteBlockService } from './note-service.js';
 import { tryUpdateNoteSize } from './utils.js';
 
 @customElement('affine-note')
@@ -22,7 +20,6 @@ export class NoteBlockComponent extends BlockElement<NoteBlockModel> {
 
   override connectedCallback() {
     super.connectedCallback();
-    registerService('affine:note', NoteBlockService);
     bindHotKey(this);
   }
