+++ packages/blocks/src/page-block/index.ts
@@ -12,6 +12,5 @@ export {
 export { readImageSize } from './edgeless/components/utils.js';
 export * from './edgeless/edgeless-page-block.js';
 export { type PageBlockModel, PageBlockSchema } from './page-model.js';
-export * from './page-service.js';
 export * from './types.js';
 export * from './utils/index.js';
