+++ packages/blocks/src/paragraph-block/index.ts
@@ -1,3 +1,2 @@
 export * from './paragraph-block.js';
 export * from './paragraph-model.js';
-export * from './paragraph-service.js';
